+++ constant/face_shape.go
@@ -8,3 +8,12 @@ var GetFaceShapeTags = map[uint64][]uint64{
 	5: {2, 6},
 	6: {1, 5},
 }
+
+var GetIDByShape = map[string]uint64{
+	"circle":   1,
+	"heart":    2,
+	"oblong":   3,
+	"oval":     4,
+	"square":   5,
+	"triangle": 6,
+}