+++ handler/auth_handler.go
@@ -14,6 +14,7 @@ func RegisterAuthRoutes(e *echo.Echo) {
 	authGroup := v1Group.Group("/auth")
 	authGroup.POST("/login", loginHandler)
 	authGroup.POST("/register-user", registerUserHandler)
+	authGroup.POST("/update-user/:id", updateUserHandler)
 	authGroup.POST("/register-partner", registerPartnerHandler)
 }
 
@@ -34,6 +35,31 @@ func loginHandler(c echo.Context) error {
 	}
 
 	return utils.ResponseJSON(c, "Login success", token, http.StatusOK)
+}
+
+func updateUserHandler(c echo.Context) error {
+	uID := utils.StrToUint64(c.Param("id"), 0)
+	if uID == 0 {
+		return utils.ResponseJSON(c, "Invalid user ID", nil, http.StatusBadRequest)
+	}
+
+	var userData datastruct.UserRegisterInput
+	err := c.Bind(&userData)
+	if err != nil {
+		return err
+	}
+
+	validationErrors := utils.ValidateStruct(userData)
+	if len(validationErrors) > 0 {
+		return utils.ResponseJSON(c, "The data is not valid", validationErrors, http.StatusBadRequest)
+	}
+
+	statusCode, err := repository.UpdateUser(userData, uID)
+	if err != nil {
+		return utils.ResponseJSON(c, err.Error(), nil, statusCode)
+	}
+
+	return utils.ResponseJSON(c, "Updated", nil, statusCode)
 }
 
 func registerUserHandler(c echo.Context) error {