+++ CHANGELOG.md
@@ -76,3 +76,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### [1.0.8] - yusufw2429@gmail.com
 #### Added
+- Add API key for integration with ML (not using jwt)