+++ CHANGELOG.md
@@ -81,3 +81,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Update deployment
 - Add resty pkg utils
 - Integrate questionnaire with ML
+- Update user profile